+++ ranlib/src/ranlib.cpp
@@ -70,7 +70,7 @@ S40:
 S50:
     w = a*exp(v);
 S60:
-    z = pow(u1,2.0)*u2;
+    z = pow(u1,2.0f)*u2;
     r = gamma*v-1.3862944;
     s = a+r-w;
 /*
@@ -129,7 +129,7 @@ S130:
 /*
      Step 3
 */
-    z = pow(u1,2.0)*u2;
+    z = pow(u1,2.0f)*u2;
     if(!(z <= 0.25)) goto S160;
     v = beta*log(u1/(1.0-u1));
     if(!(v > expmax)) goto S140;
@@ -442,7 +442,7 @@ static float gennch;
     fprintf(stderr,"Value of DF: %16.6E Value of XNONC%16.6E\n",df,xnonc);
     exit(1);
 S10:
-    gennch = genchi(df-1.0)+pow(gennor(sqrt(xnonc),1.0),2.0);
+    gennch = genchi(df-1.0)+pow(gennor(sqrt(xnonc),1.0f),2.0f);
     return gennch;
 }
 float gennf(float dfn,float dfd,float xnonc)
@@ -864,7 +864,7 @@ S140:
 /*
      INVERSE CDF LOGIC FOR MEAN LESS THAN 30
 */
-    qn = pow(q,(double)n);
+    qn = pow(q,n);
     r = p/q;
     g = r*(n+1);
 S150:
@@ -1094,7 +1094,7 @@ S70:
 */
     if(ignpoi >= 10) goto S80;
     px = -mu;
-    py = pow(mu,(double)ignpoi)/ *(fact+ignpoi);
+    py = pow(mu,ignpoi)/ *(fact+ignpoi);
     goto S110;
 S80:
 /*
